Images generate from voxel data
Inspecting the raw voxel data, I found that the top level structure was an array of
slices along the Z index. These objects were almost in the correct format for pypng
to begin with, the pixels simply had to be reduced to byte length and inverted.

I left a todo to spend more time looking at pypng itself, I suspect that
I don't have to sacrifice as much resolution as I am now to get usable output