display sorted images
it turns out that django doesn't typically handle serving
assets like images, leaving that to webservers. As this
is a puzzle and not a production environment, I didn't want
to set that up, and since the build it webserver refused
to serve images, I fell back on base64 encoding them.

Although this is technically bypassing a security mesure by allowing
static files to be servedi by django, I think it maintains the spirit of
the restriction. By not allowing files to be selected in any way other than
by selecting an image series, we still prevent access to
all information by default, thus protecting source code
and sensitive data