correctly reset the current working directory
hunted down this nasty bug, I guess the state of of the os namespace
is persistant accross requests and files!

after uploading a file, python would become unable to find files
on the local filesystem until it was restarted. This was
because the state of the os namespace had persisted being moved
to a specific image_dump directory, sever levels outside of the
project root.

other potential solutions would be to write a lexically scoped version
of chdir. Alternativly, all directories could be referenced from
an immutable global home_dir value that would be set to the
root of the project

with chdir("new_dir"):
    cwd is new_dir

cwd is old_dir