Process png creation in parallel
in clojure, list comprehension is parallelized by using nmap instead
of map. While not quite a single char change, I was able to find
something quite similiar in the multiprocessing library,
which allows one to specify a threadpool and then execute maps
in parallel with it.

After some trial and error with threadpool sizes I was able to get
the total time from file upload to completion under 23 seconds from
thr previous best of 31.

Oberving my cpu graph, it's possible that further gains could
be made parallelizing the numpy operations.